feat: Move DebugConfig and Pyroscope to common::debug
